rename GCMTools to gcmt
