--- GCMtools/__init__.py
@@ -1,4 +1,0 @@
-from .GCMtools import GCMT
-from .utils import gcm_plotting
-
-__author__ = 'Aaron David Schneider, Sven Kiefer, Robin Baeyens and Thomas Konings'
