--- GCMtools/exorad/__init__.py
@@ -1,1 +1,0 @@
-from .read_in import m_read_from_mitgcm
