+++ doc/source/installation.rst
@@ -2,7 +2,7 @@ Installation
 ============
 There are several ways to install the package.
 We recommend using anaconda.
-``GCMTools`` has been tested to work on linux and mac.
+``gcmt`` has been tested to work on linux and mac.
 
 .. note::
     Using Windows is not recommended and will not work in combination with MITgcm raw data.
@@ -24,18 +24,16 @@ Install a few dependencies (optional include the following packages: ``jupyterla
     pip install git+https://github.com/MITgcm/xmitgcm.git
 
 
-Install ``GCMTools``.
+Install ``gcmt``.
 
 .. code-block:: bash
 
-    pip install git+https://github.com/exorad/GCMtools.git
-
-We will provice anaconda and pip install methods, once ``GCMTools`` is out of beta.
+    pip install gcmt
 
 .. note::
     If you prefer to install a development version, you can use the following command:
 
     .. code-block:: bash
 
-        git clone https://github.com/exorad/GCMtools.git
+        git clone https://github.com/exorad/gcmt.git
         pip install -e GCMtools
