+++ gcmt/exorad/__init__.py
@@ -1,0 +1,1 @@
+from .read_in import m_read_from_mitgcm
